Update Board.java
Its a rainy day and so Im determined to finish this project by the end of the day as of right now I ave added offsets for the knghts, bishops, and the paws, but I still need to do the same for the queen, king and rooks. After that Im not sure how Ill handle the checks and things like that. This file is getting very long, so I am trying to put more of the if statements in functions under the piece.java file so that I dont end up with a massive file