Update Board.java
updated it so that now all pieces have their basic parameters for moving. There is still no en passant a such, but there is things like the knight work weell and no pieces are flying off one side of the board and top the other side now because i fixed it by checking the position of the piece compared to the one its supposed to move to.